+++ sageopt/tests/test_coniclifts/test_reformulators.py
@@ -111,5 +111,3 @@ class TestReformulators(unittest.TestCase):
         pass
 
 
-if __name__ == '__main__':
-    unittest.main()