+++ sageopt/tests/test_relaxations/test_sage_polynomials.py
@@ -333,7 +333,3 @@ class TestSagePolynomials(unittest.TestCase):
         gap = abs(f(x_star) - opt)
         assert gap < 1e-4
         pass
-
-
-if __name__ == '__main__':
-    unittest.main()